pr-playerArea
